raspberrypi/alarm_transmitter/requirements.txt
python-dotenv==1.0.1
loguru==0.7.3
pyserial==3.5
RPi.GPIO==0.7.1
rpi-lgpio==0.6